pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.7.4",
    "beautifulsoup4>=4.12.0",
    "boto3>=1.26.0",
    "fake-useragent>=1.1.0",
    "openai>=0.27.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "streamlit-tags>=1.2.0",
    "streamlit>=1.0.0",
    "streamlit-option-menu>=0.3.0",
    "tenacity>=8.2.0",
    "google>=3.0.0",
    "email-validator>=2.0.0",
    "plotly>=5.14.0",
    "requests>=2.25.1",
    "psycopg-binary>=2.9.2",
    "streamlit-lottie>=0.0.3",
    "streamlit-timeline>=0.0.2",
    "phonenumbers>=8.13.0",
    "psycopg2-binary>=2.9.0",
    "botocore>=1.29.0",
    "gradio-client>=0.2.0",
    "transformers>=4.28.0",
    "typing-extensions>=4.0.0",
    "urllib3>=1.26.0,<2.0.0",
    "requests-toolbelt>=1.0.0",
    "streamlit-aggrid>=0.3.0",
    "black>=22.0.0",
    "flask>=2.0.0",
    "gradio>=3.0.0",
    "pydantic>=1.10.0",
    "faiss-cpu>=1.7.0",
    "loguru>=0.7.0",
    "pytest>=6.0.0",
    "rich>=10.0.0",
    "huggingface-hub>=0.13.0",
    "redis>=4.5.0",
    "streamlit-extras>=0.2.0",
    "pyyaml>=6.0.0",
    "apscheduler>=3.10.0",
    "watchdog>=3.0.0",
    "debugpy>=1.6.0",
    "trafilatura>=2.0.0",
    "psutil>=6.1.1",
    "importlib>=1.0.4",
    "google-search>=1.1.1",
    "googlesearch-python>=1.2.5",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.25.2",
]
